prune Shard and reorganize
