+++ tsdb/engine.go
@@ -123,10 +123,11 @@ func RegisteredEngines() []string {
 
 // NewEngine returns an instance of an engine based on its format.
 // If the path does not exist then the DefaultFormat is used.
-func NewEngine(id uint64, i Index, path string, walPath string, sfile *SeriesFile, options EngineOptions) (Engine, error) {
+func NewEngine(id uint64, i Index, path string, sfile *SeriesFile, options EngineOptions) (Engine, error) {
 	// Create a new engine
 	if _, err := os.Stat(path); os.IsNotExist(err) {
-		engine := newEngineFuncs[options.EngineVersion](id, i, path, walPath, sfile, options)
+		// TODO(jeff): remove walPath argument
+		engine := newEngineFuncs[options.EngineVersion](id, i, path, "", sfile, options)
 		if options.OnNewEngine != nil {
 			options.OnNewEngine(engine)
 		}
@@ -149,7 +150,8 @@ func NewEngine(id uint64, i Index, path string, walPath string, sfile *SeriesFil
 		return nil, fmt.Errorf("invalid engine format: %q", format)
 	}
 
-	engine := fn(id, i, path, walPath, sfile, options)
+	// TODO(jeff): remove walPath argument
+	engine := fn(id, i, path, "", sfile, options)
 	if options.OnNewEngine != nil {
 		options.OnNewEngine(engine)
 	}
