+++ tsdb/errors.go
@@ -1,0 +1,77 @@
+package tsdb
+
+import (
+	"errors"
+	"fmt"
+)
+
+var (
+	// ErrFieldOverflow is returned when too many fields are created on a measurement.
+	ErrFieldOverflow = errors.New("field overflow")
+
+	// ErrFieldTypeConflict is returned when a new field already exists with a different type.
+	ErrFieldTypeConflict = errors.New("field type conflict")
+
+	// ErrFieldNotFound is returned when a field cannot be found.
+	ErrFieldNotFound = errors.New("field not found")
+
+	// ErrFieldUnmappedID is returned when the system is presented, during decode, with a field ID
+	// there is no mapping for.
+	ErrFieldUnmappedID = errors.New("field ID not mapped")
+
+	// ErrEngineClosed is returned when a caller attempts indirectly to
+	// access the shard's underlying engine.
+	ErrEngineClosed = errors.New("engine is closed")
+
+	// ErrShardDisabled is returned when a the shard is not available for
+	// queries or writes.
+	ErrShardDisabled = errors.New("shard is disabled")
+
+	// ErrUnknownFieldsFormat is returned when the fields index file is not identifiable by
+	// the file's magic number.
+	ErrUnknownFieldsFormat = errors.New("unknown field index format")
+
+	// ErrUnknownFieldType is returned when the type of a field cannot be determined.
+	ErrUnknownFieldType = errors.New("unknown field type")
+
+	// ErrShardNotIdle is returned when an operation requring the shard to be idle/cold is
+	// attempted on a hot shard.
+	ErrShardNotIdle = errors.New("shard not idle")
+
+	// fieldsIndexMagicNumber is the file magic number for the fields index file.
+	fieldsIndexMagicNumber = []byte{0, 6, 1, 3}
+)
+
+// A ShardError implements the error interface, and contains extra
+// context about the shard that generated the error.
+type ShardError struct {
+	id  uint64
+	Err error
+}
+
+// NewShardError returns a new ShardError.
+func NewShardError(id uint64, err error) error {
+	if err == nil {
+		return nil
+	}
+	return ShardError{id: id, Err: err}
+}
+
+// Error returns the string representation of the error, to satisfy the error interface.
+func (e ShardError) Error() string {
+	return fmt.Sprintf("[shard %d] %s", e.id, e.Err)
+}
+
+// PartialWriteError indicates a write request could only write a portion of the
+// requested values.
+type PartialWriteError struct {
+	Reason  string
+	Dropped int
+
+	// A sorted slice of series keys that were dropped.
+	DroppedKeys [][]byte
+}
+
+func (e PartialWriteError) Error() string {
+	return fmt.Sprintf("partial write: %s dropped=%d", e.Reason, e.Dropped)
+}
