+++ tsdb/index.go
@@ -2672,7 +2672,7 @@ func RegisteredIndexes() []string {
 
 // NewIndex returns an instance of an index based on its format.
 // If the path does not exist then the DefaultFormat is used.
-func NewIndex(id uint64, database, path string, seriesIDSet *SeriesIDSet, sfile *SeriesFile, options EngineOptions) (Index, error) {
+func NewIndex(id uint64, path string, seriesIDSet *SeriesIDSet, sfile *SeriesFile, options EngineOptions) (Index, error) {
 	format := options.IndexVersion
 
 	// Use default format unless existing directory exists.
@@ -2690,11 +2690,13 @@ func NewIndex(id uint64, database, path string, seriesIDSet *SeriesIDSet, sfile 
 	if fn == nil {
 		return nil, fmt.Errorf("invalid index format: %q", format)
 	}
-	return fn(id, database, path, seriesIDSet, sfile, options), nil
+
+	// TODO(jeff): remove database argument
+	return fn(id, "", path, seriesIDSet, sfile, options), nil
 }
 
-func MustOpenIndex(id uint64, database, path string, seriesIDSet *SeriesIDSet, sfile *SeriesFile, options EngineOptions) Index {
-	idx, err := NewIndex(id, database, path, seriesIDSet, sfile, options)
+func MustOpenIndex(id uint64, path string, seriesIDSet *SeriesIDSet, sfile *SeriesFile, options EngineOptions) Index {
+	idx, err := NewIndex(id, path, seriesIDSet, sfile, options)
 	if err != nil {
 		panic(err)
 	} else if err := idx.Open(); err != nil {
